handle git branches in target path
This change adds handling for the magic path element GIT_BRANCH to the remote path. It will be re-evaluated with each sync operation. Meaning if you change branches while synchrotron is running, it will sync the NEW branch to it's own dir on the server, and not crush the one you had previously been syncing.